requirements.txt
alembic==1.11.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
databricks-cli==0.17.7
docker==6.1.2
entrypoints==0.4
Flask==2.3.2
fonttools==4.39.4
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kafka-python==2.0.2
kiwisolver==1.4.4
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
mlflow==2.3.2
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==4.23.0
pyarrow==11.0.0
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
querystring-parser==1.2.4
requests==2.30.0
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.23.0
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.13
sqlparse==0.4.4
tabulate==0.9.0
threadpoolctl==3.1.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
websocket-client==1.5.1
Werkzeug==2.3.4
xgboost==1.7.5
zipp==3.15.0
